refactor: extract localStorage mock setup to utility function
Encapsulate localStorage mock setup/teardown logic in a reusable utility
function to eliminate duplication and improve maintainability.

Changes:
- Add setupLocalStorageMock() utility to test-utils.tsx
- Update both test files to use the new utility

This addresses code review feedback to reduce duplication between test files
and makes the pattern easier to reuse across the test suite.

Related: #257
