+++ app/__tests__/hooks/useTodos.softdelete.test.ts
@@ -1,30 +1,12 @@
 import { renderHook, act } from '@testing-library/react';
 import { useTodos } from '../../hooks/useTodos';
-import { mockLocalStorage } from '../utils/test-utils';
+import { setupLocalStorageMock } from '../utils/test-utils';
 
-// Mock localStorage
-const mockStorage = mockLocalStorage();
-const originalLocalStorage = window.localStorage;
-
-Object.defineProperty(window, 'localStorage', {
-  value: mockStorage,
-  writable: true,
-  configurable: true,
-});
+// Setup localStorage mock with automatic cleanup
+const { mockStorage, restoreLocalStorage } = setupLocalStorageMock();
 
 describe('useTodos Hook - Soft Delete Functionality', () => {
-  afterAll(() => {
-    // Restore original localStorage to maintain test isolation
-    if (originalLocalStorage) {
-      Object.defineProperty(window, 'localStorage', {
-        value: originalLocalStorage,
-        writable: true,
-        configurable: true,
-      });
-    } else {
-      delete (window as { localStorage?: Storage }).localStorage;
-    }
-  });
+  afterAll(restoreLocalStorage);
 
   beforeEach(() => {
     // Clear localStorage before each test
